samples/snippets/requirements-test.txt
pytest==6.2.5
google-cloud-testutils==1.3.0
google-api-core==2.0.1
google-cloud-storage